backend/pyproject.toml
[project]
name = "eero-ui-backend"
version = "1.0.0"
description = "FastAPI backend for the Eero UI Dashboard"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "eero-client @ git+https://github.com/fulviofreitas/eero-client.git",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.26.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
